+++ dphonebook/lib/providers/__init__.py
@@ -1,5 +1,6 @@
 from dphonebook.lib.providers.dummy_example_com import DummyExampleCom
 from dphonebook.lib.providers.receive_sms_co import ReceiveSmsCo
+from dphonebook.lib.providers.receive_sms_free_cc import ReceiveSmsFreeCc
 from dphonebook.lib.providers.receive_sms_online_info import ReceiveSmsOnlineInfo
 from dphonebook.lib.providers.receive_sms_org import ReceiveSmsOrg
 from dphonebook.lib.providers.receive_smss import ReceiveSmss
@@ -12,5 +13,6 @@ number_provider_classes = [
     ReceiveSmsCo,
     ReceiveSmsOnlineInfo,
     ReceiveSmsOrg,
-    SmsOnlineCo
+    SmsOnlineCo,
+    ReceiveSmsFreeCc
 ]